Header Changes
Touched up the header, made it so that only the links to the sign-in page are in the header for all of the current pages except the dashboard. Other links are on the dashboard, but they all take you to the login page seeing as the edit profile and add file pages aren't implemented yet.